add --no-drop option
by default, when all voices are playing, we will not play a new note
if it's not higher or lower than any playing note. this option
forces the player to preempt the middle note for the new one.

Pachelbel's Canon sounds better with this option.
Aliens Ate My Babysitter sounds worse.